Extracting 'infinity extractor' from kalibro range class to a module  - Applaying Single responsability principle from SOLID
